Manage clicking outside of a stage. Minor Refactor.
