+++ illum/package.json
@@ -12,7 +12,8 @@
     "next": "^15.1.7",
     "react": "^19.0.0",
     "react-click-outside": "^3.0.1",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-outside-click-handler": "^1.3.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
@@ -20,6 +21,7 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "@types/react-onclickoutside": "^6.7.10",
+    "@types/react-outside-click-handler": "^1.3.4",
     "eslint": "^9",
     "eslint-config-next": "15.1.7",
     "postcss": "^8",
